Added visit tracking for links
I added a visits column to the links table for counting visits to a
link, and I added a show action to the links controller and routes for
incrementing the visits count then redirecting to the given link. This
way I can track how many times a link is visited.