Changed feed model to fetch on a job delay
I moved feed fetching into a delayed job worker, and I added a frequency
column to the feed model for calculating how often rss feeds should be
queried. I added a rudimentary algorithm for calculating the frequency
by dividing the frequency when new stories are present and doubling when
none are present.